+++ app/assets/javascripts/admin/admin_rest_api.js
@@ -51,6 +51,7 @@ import type { QueryObject } from "admin/task/task_search_form";
 import type { NewTask, TaskCreationResponse } from "admin/task/task_create_bulk_view";
 import type { DatasetConfiguration } from "oxalis/store";
 import type { RequestOptions } from "libs/request";
+import type { Versions } from "oxalis/view/version_view";
 
 const MAX_SERVER_ITEMS_PER_RESPONSE = 1000;
 
@@ -71,10 +72,12 @@ function requestUserToken(): Promise<string> {
     return tokenRequestPromise;
   }
 
-  tokenRequestPromise = Request.receiveJSON("/api/userToken/generate").then(tokenObj => {
-    tokenRequestPromise = null;
-    return tokenObj.token;
-  });
+  tokenRequestPromise = Request.receiveJSON("/api/userToken/generate", { method: "POST" }).then(
+    tokenObj => {
+      tokenRequestPromise = null;
+      return tokenObj.token;
+    },
+  );
 
   return tokenRequestPromise;
 }
@@ -283,6 +286,7 @@ export async function increaseProjectTaskInstances(
 ): Promise<APIProjectWithAssignments> {
   const project = await Request.receiveJSON(
     `/api/projects/${projectName}/incrementEachTasksInstances?delta=${delta}`,
+    { method: "PATCH" },
   );
   return transformProject(project);
 }
@@ -318,12 +322,16 @@ export function updateProject(
 }
 
 export async function pauseProject(projectName: string): Promise<APIProject> {
-  const project = await Request.receiveJSON(`/api/projects/${projectName}/pause`);
+  const project = await Request.receiveJSON(`/api/projects/${projectName}/pause`, {
+    method: "PATCH",
+  });
   return transformProject(project);
 }
 
 export async function resumeProject(projectName: string): Promise<APIProject> {
-  const project = await Request.receiveJSON(`/api/projects/${projectName}/resume`);
+  const project = await Request.receiveJSON(`/api/projects/${projectName}/resume`, {
+    method: "PATCH",
+  });
   return transformProject(project);
 }
 
@@ -333,7 +341,7 @@ export function peekNextTasks(): Promise<?APITask> {
 }
 
 export function requestTask(): Promise<APIAnnotationWithTask> {
-  return Request.receiveJSON("/api/user/tasks/request");
+  return Request.receiveJSON("/api/user/tasks/request", { method: "POST" });
 }
 
 export function getAnnotationsForTask(taskId: string): Promise<Array<APIAnnotation>> {
@@ -404,6 +412,7 @@ export function finishTask(annotationId: string): Promise<APIAnnotation> {
 
 export function transferTask(annotationId: string, userId: string): Promise<APIAnnotation> {
   return Request.sendJSONReceiveJSON(`/api/annotations/Task/${annotationId}/transfer`, {
+    method: "PATCH",
     data: {
       userId,
     },
@@ -446,7 +455,9 @@ export function reOpenAnnotation(
   annotationId: string,
   annotationType: APITracingType,
 ): Promise<APIAnnotation> {
-  return Request.receiveJSON(`/api/annotations/${annotationType}/${annotationId}/reopen`);
+  return Request.receiveJSON(`/api/annotations/${annotationType}/${annotationId}/reopen`, {
+    method: "PATCH",
+  });
 }
 
 export type EditableAnnotation = {
@@ -463,6 +474,7 @@ export function editAnnotation(
 ): Promise<void> {
   return Request.sendJSONReceiveJSON(`/api/annotations/${annotationType}/${annotationId}/edit`, {
     data,
+    method: "PATCH",
   });
 }
 
@@ -470,14 +482,18 @@ export function finishAnnotation(
   annotationId: string,
   annotationType: APITracingType,
 ): Promise<APIAnnotation> {
-  return Request.receiveJSON(`/api/annotations/${annotationType}/${annotationId}/finish`);
+  return Request.receiveJSON(`/api/annotations/${annotationType}/${annotationId}/finish`, {
+    method: "PATCH",
+  });
 }
 
 export function resetAnnotation(
   annotationId: string,
   annotationType: APITracingType,
 ): Promise<APIAnnotation> {
-  return Request.receiveJSON(`/api/annotations/${annotationType}/${annotationId}/reset`);
+  return Request.receiveJSON(`/api/annotations/${annotationType}/${annotationId}/reset`, {
+    method: "PUT",
+  });
 }
 
 export function deleteAnnotation(
@@ -493,6 +509,7 @@ export function finishAllAnnotations(
   selectedAnnotationIds: Array<string>,
 ): Promise<{ messages: Array<Message> }> {
   return Request.sendJSONReceiveJSON("/api/annotations/Explorational/finish", {
+    method: "PATCH",
     data: {
       annotations: selectedAnnotationIds,
     },
@@ -504,7 +521,7 @@ export function copyAnnotationToUserAccount(
   tracingType: APITracingType,
 ): Promise<APIAnnotation> {
   const url = `/api/annotations/${tracingType}/${annotationId}/duplicate`;
-  return Request.receiveJSON(url);
+  return Request.receiveJSON(url, { method: "POST" });
 }
 
 export function getAnnotationInformation(
@@ -530,11 +547,11 @@ export function createExplorational(
 
 export async function getTracingForAnnotations(
   annotation: APIAnnotation,
-  version?: number,
+  versions?: Versions = {},
 ): Promise<HybridServerTracing> {
   const [_skeleton, _volume] = await Promise.all([
-    getTracingForAnnotationType(annotation, "skeleton", version),
-    getTracingForAnnotationType(annotation, "volume", version),
+    getTracingForAnnotationType(annotation, "skeleton", versions.skeleton),
+    getTracingForAnnotationType(annotation, "volume", versions.volume),
   ]);
 
   const skeleton = ((_skeleton: any): ?ServerSkeletonTracing);
@@ -549,7 +566,7 @@ export async function getTracingForAnnotations(
 export async function getTracingForAnnotationType(
   annotation: APIAnnotation,
   tracingType: "skeleton" | "volume",
-  version?: number,
+  version?: ?number,
 ): Promise<?ServerTracing> {
   const tracingId = annotation.tracing[tracingType];
   if (!tracingId) {
@@ -646,6 +663,7 @@ export function updateDataset(datasetId: APIDatasetId, dataset: APIDataset): Pro
   return Request.sendJSONReceiveJSON(
     `/api/datasets/${datasetId.owningOrganization}/${datasetId.name}`,
     {
+      method: "PATCH",
       data: dataset,
     },
   );
@@ -654,6 +672,19 @@ export function updateDataset(datasetId: APIDatasetId, dataset: APIDataset): Pro
 export function getDatasetConfiguration(datasetId: APIDatasetId): Promise<Object> {
   return Request.receiveJSON(
     `/api/dataSetConfigurations/${datasetId.owningOrganization}/${datasetId.name}`,
+  );
+}
+
+export function updateDatasetConfiguration(
+  datasetId: APIDatasetId,
+  datasetConfig: DatasetConfiguration,
+): Object {
+  return Request.sendJSONReceiveJSON(
+    `/api/dataSetConfigurations/${datasetId.owningOrganization}/${datasetId.name}`,
+    {
+      method: "PUT",
+      data: datasetConfig,
+    },
   );
 }
 
@@ -672,6 +703,7 @@ export function updateDatasetDefaultConfiguration(
   return Request.sendJSONReceiveJSON(
     `/api/dataSetConfigurations/default/${datasetId.owningOrganization}/${datasetId.name}`,
     {
+      method: "PUT",
       data: datasetConfiguration,
     },
   );
@@ -733,6 +765,7 @@ export function updateDatasetTeams(
   return Request.sendJSONReceiveJSON(
     `/api/datasets/${datasetId.owningOrganization}/${datasetId.name}/teams`,
     {
+      method: "PATCH",
       data: newTeams,
     },
   );
@@ -797,6 +830,13 @@ export function getActiveUser(options: Object = {}): Promise<APIUser> {
 
 export function getUserConfiguration(): Object {
   return Request.receiveJSON("/api/user/userConfiguration");
+}
+
+export function updateUserConfiguration(userConfiguration: Object): Object {
+  return Request.sendJSONReceiveJSON("/api/user/userConfiguration", {
+    method: "PUT",
+    data: userConfiguration,
+  });
 }
 
 // ### TimeTracking