+++ webknossos-datastore/app/com/scalableminds/webknossos/datastore/DataStoreModule.scala
@@ -15,6 +15,7 @@ class DataStoreModule(environment: Environment, configuration: Configuration) ex
 
   def configure() = {
     bind(classOf[DataStoreConfig]).asEagerSingleton()
+    bind(classOf[BaseDirService]).asEagerSingleton()
     bind(classOf[AccessTokenService]).asEagerSingleton()
     bind(classOf[ActorSystem]).annotatedWith(Names.named("webknossos-datastore")).toInstance(system)
     bind(classOf[BinaryDataService]).asEagerSingleton()